% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{addItemToFeed}
\alias{addItemToFeed}
\title{Add items to your feed}
\usage{
addItemToFeed(mtoken = getMonzoToken(), accountId = getDefaultAccountId(),
  title = NULL, itembody = NULL, image_url = NULL,
  background_colour = NULL, title_colour = NULL, body_colour = NULL)
}
\arguments{
\item{mtoken}{The Monzo API token}

\item{accountId}{The id of the account you're requesting transactions from}

\item{title}{The title of the item to be displayed in the feed}

\item{image_url}{The URL of the image/icon to be displayed alongside the item in the feed}

\item{background_colour}{The hex value of the colour of the background for the feed item}

\item{title_colour}{The hex value of the colour of the title text for the feed item}

\item{body_colour}{The hex value of the colour of the body text for the feed item}

\item{body}{The body text of the item to be displayed in the feed}
}
\description{
Add item to feed
}
\keyword{feed}
