% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pots.R
\name{depositIntoPot}
\alias{depositIntoPot}
\title{Deposit in a pot}
\usage{
depositIntoPot(mtoken = getMonzoToken(), accountId = NULL, potId = NULL,
  amount = NULL, dedupeId = NULL)
}
\arguments{
\item{mtoken}{The Monzo API token}

\item{accountId}{The source id of the account you're transferring from}

\item{potId}{The id of the pot}

\item{amount}{The amount to deposit in the pot}

\item{dedupeId}{The id required to ensure the deposit is unique}
}
\description{
Deposit in a pot
}
\keyword{pots}
