% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{getAccounts}
\alias{getAccounts}
\title{Retrieving and managing accounts}
\usage{
getAccounts(mtoken = getMonzoToken())
}
\arguments{
\item{mtoken}{the Monzo API auth token}
}
\description{
These functions are intended to retrieve Accounts
}
\keyword{accounts}
