% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{getDefaultAccountId}
\alias{getDefaultAccountId}
\title{Retrieve the first account ID for a valid account}
\usage{
getDefaultAccountId(mtoken = getMonzoToken())
}
\arguments{
\item{mtoken}{The Monzo API auth token}
}
\description{
Retrieve the first account ID for a valid account
}
\keyword{accountId}
