% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{getMonzoTransactions}
\alias{getMonzoTransactions}
\title{Retrieving transactions and including items in the feed}
\usage{
getMonzoTransactions(mtoken = getMonzoToken(), accountId = NULL)
}
\arguments{
\item{mtoken}{The Monzo API token}

\item{accountId}{The id of the account you're requesting transactions from}
}
\description{
These functions are intended to create and retrieve transactions
}
\keyword{transactions}
