% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pots.R
\name{getPots}
\alias{getPots}
\title{Retrieve, withdraw from, and deposit into pots}
\usage{
getPots(mtoken = getMonzoToken(), accountId = NULL)
}
\arguments{
\item{mtoken}{The Monzo API token}

\item{accountId}{The id of the account you're requesting transactions from}
}
\description{
Retrieve pots, defaulting to the first valid account if no account ID is supplied
}
\keyword{pots}
