% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{getTransaction}
\alias{getTransaction}
\title{Retrieve a specific transaction, defaulting to the first valid account if no account ID is supplied}
\usage{
getTransaction(mtoken = getMonzoToken(), accountId = NULL,
  transactionId = NULL, expand = "merchant", parse = TRUE)
}
\arguments{
\item{mtoken}{The Monzo API token}

\item{accountId}{The id of the account that you're requesting transactions from}

\item{transactionId}{The id of the transaction being requested}

\item{expand}{A string dictating what should be expanded within the responses. Defaults to 'merchant'}

\item{parse}{Boolean dictating whether to coerce response into a neater dataframe. Defaults to true}
}
\description{
Retrieve a specific transaction, defaulting to the first valid account if no account ID is supplied
}
\keyword{transactions}
