% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{getTransactions}
\alias{getTransactions}
\title{Retrieving transactions and including items in the feed}
\usage{
getTransactions(mtoken = getMonzoToken(), accountId = NULL, since = NULL,
  before = NULL, parse = TRUE, expand = NULL)
}
\arguments{
\item{mtoken}{The Monzo API token}

\item{accountId}{The id of the account you're requesting transactions from}

\item{since}{An RFC 3339 encoded-timestamp}

\item{before}{An RFC 3339 encoded-timestamp}

\item{parse}{Boolean dictating whether to coerce response into a neater dataframe. Defaults to true}

\item{expand}{A string dictating what should be expanded within the responses. Defaults to null}
}
\description{
Retrieve transactions, defaulting to the first valid account if no account ID is supplied
}
\keyword{transactions}
